Split out ProjectLocale & ProjectLocaleQuerySet
